refactor: address CodeRabbit review suggestions
- Fix TOCTOU race: insert atomically then send, instead of check-then-send
- Add err to tracing::instrument on Result-returning DB functions
- Move tests to separate test.rs file per project conventions
