--- rust/cloud-storage/macro_db_client/.sqlx/query-1435acbba365796ddcb6ddbef86f4ebf664cffc44082996f6b60a60373c28f75.json
@@ -1,17 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n                INSERT INTO entity_properties (id, entity_id, entity_type, property_definition_id, values)\n                SELECT \n                    u.id,\n                    u.entity_id,\n                    'TASK'::property_entity_type,\n                    u.property_definition_id,\n                    u.values\n                FROM UNNEST(\n                    $1::UUID[],\n                    $2::TEXT[],\n                    $3::UUID[],\n                    $4::JSONB[]\n                ) AS u(id, entity_id, property_definition_id, values)\n                ON CONFLICT (entity_id, entity_type, property_definition_id)\n                DO UPDATE SET \n                    values = EXCLUDED.values,\n                    updated_at = NOW()\n                ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "UuidArray",
-        "TextArray",
-        "UuidArray",
-        "JsonbArray"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "1435acbba365796ddcb6ddbef86f4ebf664cffc44082996f6b60a60373c28f75"
-}
