--- rust/cloud-storage/macro_db_client/.sqlx/query-37b5168f739c490a3373c65eab35f247a34af0a8e659cbad61cb99e3094e5a60.json
@@ -1,23 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT EXISTS (\n                SELECT 1\n                FROM public.email_threads t\n                JOIN public.email_links l ON l.id = t.link_id\n                WHERE t.id = $1::uuid\n                  AND l.macro_id = $2\n            ) AS \"exists!\"\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "exists!",
-        "type_info": "Bool"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Text"
-      ]
-    },
-    "nullable": [
-      null
-    ]
-  },
-  "hash": "37b5168f739c490a3373c65eab35f247a34af0a8e659cbad61cb99e3094e5a60"
-}
