--- rust/cloud-storage/macro_db_client/.sqlx/query-63492b018803ce26db5891964a0f52f37d2f54bd8243265c6561f4423af8617c.json
@@ -1,16 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            INSERT INTO entity_properties (id, entity_id, entity_type, property_definition_id, values)\n            SELECT \n                u.id,\n                u.entity_id,\n                'TASK'::property_entity_type,\n                u.property_definition_id,\n                NULL\n            FROM UNNEST(\n                $1::UUID[],\n                $2::TEXT[],\n                $3::UUID[]\n            ) AS u(id, entity_id, property_definition_id)\n            ON CONFLICT (entity_id, entity_type, property_definition_id)\n            DO NOTHING\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "UuidArray",
-        "TextArray",
-        "UuidArray"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "63492b018803ce26db5891964a0f52f37d2f54bd8243265c6561f4423af8617c"
-}
