--- rust/cloud-storage/macro_db_client/.sqlx/query-7617f808df94ffe77c8195a911ee6b230a21fe9935e099d1310d244248ca253a.json
@@ -1,23 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n        SELECT c.id\n        FROM comms_channels c\n        INNER JOIN comms_channel_participants cp ON cp.channel_id = c.id\n        WHERE cp.user_id = $1 AND cp.left_at IS NULL\n        AND c.id = ANY($2::uuid[])\n        ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Text",
-        "UuidArray"
-      ]
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "7617f808df94ffe77c8195a911ee6b230a21fe9935e099d1310d244248ca253a"
-}
