--- rust/cloud-storage/macro_db_client/.sqlx/query-8d21bf0aa4c20589a55cf5777aa9c658fdacea4d0d0c7070aded12cccd373911.json
@@ -1,16 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            UPDATE entity_properties\n            SET values = $2\n            WHERE entity_id = $1\n            AND property_definition_id = $3\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Text",
-        "Jsonb",
-        "Uuid"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "8d21bf0aa4c20589a55cf5777aa9c658fdacea4d0d0c7070aded12cccd373911"
-}
