--- rust/cloud-storage/macro_db_client/.sqlx/query-9cd40fbb1ade51e9e0d49792839d1a363cada017252a6f653c2272608189bcc4.json
@@ -1,28 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            SELECT property_definition_id, values\n            FROM entity_properties\n            WHERE entity_id = $1\n              AND entity_type = 'TASK'\n            ",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "property_definition_id",
-        "type_info": "Uuid"
-      },
-      {
-        "ordinal": 1,
-        "name": "values",
-        "type_info": "Jsonb"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Text"
-      ]
-    },
-    "nullable": [
-      false,
-      true
-    ]
-  },
-  "hash": "9cd40fbb1ade51e9e0d49792839d1a363cada017252a6f653c2272608189bcc4"
-}
