--- rust/cloud-storage/macro_db_client/.sqlx/query-f2d5d7aea9918978cd69c14f8afe24156c2fa387e001ee8eaf57e4a70a106069.json
@@ -1,18 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "\n            INSERT INTO entity_properties (id, entity_id, entity_type, property_definition_id, values)\n            SELECT \n                u.id,\n                u.entity_id,\n                u.entity_type::property_entity_type,\n                u.property_definition_id,\n                u.values\n            FROM UNNEST(\n                $1::UUID[],\n                $2::TEXT[],\n                $3::TEXT[],\n                $4::UUID[],\n                $5::JSONB[]\n            ) AS u(id, entity_id, entity_type, property_definition_id, values)\n            ON CONFLICT (entity_id, entity_type, property_definition_id)\n            DO UPDATE SET \n                values = EXCLUDED.values,\n                updated_at = NOW()\n            ",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "UuidArray",
-        "TextArray",
-        "TextArray",
-        "UuidArray",
-        "JsonbArray"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "f2d5d7aea9918978cd69c14f8afe24156c2fa387e001ee8eaf57e4a70a106069"
-}
